pyproject.toml
[project]
name = "gai-prompt-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docarray>=0.40.0",
    "langchain-ollama>=0.2.2",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "pypdf>=5.1.0",
    "faiss-cpu>=1.9.0.post1",
    "xmltodict>=0.14.2",
    "langchain-openai>=0.3.1",
    "python-dotenv>=1.0.1",
    "langchain-chroma>=0.2.0",
    "pyright>=1.1.392.post0",
]
